+++ app/code/Magento/GoogleAdwords/Helper/Data.php
@@ -109,12 +109,15 @@ class Data extends \Magento\App\Helper\AbstractHelper
      */
     public function isGoogleAdwordsActive()
     {
-        return $this->_storeConfig->isSetFlag(self::XML_PATH_ACTIVE, \Magento\Store\Model\ScopeInterface::SCOPE_STORE)
-            && $this->getConversionId()
-            && $this->getConversionLanguage()
-            && $this->getConversionFormat()
-            && $this->getConversionColor()
-            && $this->getConversionLabel();
+        return $this->_storeConfig->isSetFlag(
+            self::XML_PATH_ACTIVE,
+            \Magento\Store\Model\ScopeInterface::SCOPE_STORE
+        ) &&
+            $this->getConversionId() &&
+            $this->getConversionLanguage() &&
+            $this->getConversionFormat() &&
+            $this->getConversionColor() &&
+            $this->getConversionLabel();
     }
 
     /**
@@ -170,7 +173,10 @@ class Data extends \Magento\App\Helper\AbstractHelper
      */
     public function getConversionId()
     {
-        return (int)$this->_storeConfig->getValue(self::XML_PATH_CONVERSION_ID, \Magento\Store\Model\ScopeInterface::SCOPE_STORE);
+        return (int)$this->_storeConfig->getValue(
+            self::XML_PATH_CONVERSION_ID,
+            \Magento\Store\Model\ScopeInterface::SCOPE_STORE
+        );
     }
 
     /**
@@ -180,7 +186,10 @@ class Data extends \Magento\App\Helper\AbstractHelper
      */
     public function getConversionLanguage()
     {
-        return $this->_storeConfig->getValue(self::XML_PATH_CONVERSION_LANGUAGE, \Magento\Store\Model\ScopeInterface::SCOPE_STORE);
+        return $this->_storeConfig->getValue(
+            self::XML_PATH_CONVERSION_LANGUAGE,
+            \Magento\Store\Model\ScopeInterface::SCOPE_STORE
+        );
     }
 
     /**
@@ -190,7 +199,10 @@ class Data extends \Magento\App\Helper\AbstractHelper
      */
     public function getConversionFormat()
     {
-        return $this->_storeConfig->getValue(self::XML_PATH_CONVERSION_FORMAT, \Magento\Store\Model\ScopeInterface::SCOPE_STORE);
+        return $this->_storeConfig->getValue(
+            self::XML_PATH_CONVERSION_FORMAT,
+            \Magento\Store\Model\ScopeInterface::SCOPE_STORE
+        );
     }
 
     /**
@@ -200,7 +212,10 @@ class Data extends \Magento\App\Helper\AbstractHelper
      */
     public function getConversionColor()
     {
-        return $this->_storeConfig->getValue(self::XML_PATH_CONVERSION_COLOR, \Magento\Store\Model\ScopeInterface::SCOPE_STORE);
+        return $this->_storeConfig->getValue(
+            self::XML_PATH_CONVERSION_COLOR,
+            \Magento\Store\Model\ScopeInterface::SCOPE_STORE
+        );
     }
 
     /**
@@ -210,7 +225,10 @@ class Data extends \Magento\App\Helper\AbstractHelper
      */
     public function getConversionLabel()
     {
-        return $this->_storeConfig->getValue(self::XML_PATH_CONVERSION_LABEL, \Magento\Store\Model\ScopeInterface::SCOPE_STORE);
+        return $this->_storeConfig->getValue(
+            self::XML_PATH_CONVERSION_LABEL,
+            \Magento\Store\Model\ScopeInterface::SCOPE_STORE
+        );
     }
 
     /**
@@ -220,7 +238,10 @@ class Data extends \Magento\App\Helper\AbstractHelper
      */
     public function getConversionValueType()
     {
-        return $this->_storeConfig->getValue(self::XML_PATH_CONVERSION_VALUE_TYPE, \Magento\Store\Model\ScopeInterface::SCOPE_STORE);
+        return $this->_storeConfig->getValue(
+            self::XML_PATH_CONVERSION_VALUE_TYPE,
+            \Magento\Store\Model\ScopeInterface::SCOPE_STORE
+        );
     }
 
     /**
@@ -240,7 +261,10 @@ class Data extends \Magento\App\Helper\AbstractHelper
      */
     public function getConversionValueConstant()
     {
-        return (float)$this->_storeConfig->getValue(self::XML_PATH_CONVERSION_VALUE, \Magento\Store\Model\ScopeInterface::SCOPE_STORE);
+        return (double)$this->_storeConfig->getValue(
+            self::XML_PATH_CONVERSION_VALUE,
+            \Magento\Store\Model\ScopeInterface::SCOPE_STORE
+        );
     }
 
     /**
