+++ app/code/Magento/Rss/Block/Catalog/NotifyStock.php
@@ -80,9 +80,10 @@ class NotifyStock extends \Magento\Backend\Block\AbstractBlock
             array('title' => $title, 'description' => $title, 'link' => $newUrl, 'charset' => 'UTF-8')
         );
 
-        $globalNotifyStockQty = (float)$this->_storeConfig->getValue(
-            \Magento\CatalogInventory\Model\Stock\Item::XML_PATH_NOTIFY_STOCK_QTY
-        , \Magento\Store\Model\ScopeInterface::SCOPE_STORE);
+        $globalNotifyStockQty = (double)$this->_storeConfig->getValue(
+            \Magento\CatalogInventory\Model\Stock\Item::XML_PATH_NOTIFY_STOCK_QTY,
+            \Magento\Store\Model\ScopeInterface::SCOPE_STORE
+        );
         /* @var $product \Magento\Catalog\Model\Product */
         $product = $this->_productFactory->create();
         /* @var $collection \Magento\Catalog\Model\Resource\Product\Collection */
